update the routers
